commandHauls: extract ProgressMonitor to module
